Drop lib/config.py and config.json
Rationale:

1. Most of these variables apply in a single location in the
notebook. If we follow the same thinking as the per-feature
refactoring, it's better to have the settings closer to the code it
affects, instead of declared at the top of the notebook.
For instance:
    layers.Dense(64)(x)
is more explicit than:
    layers.Dense(model_config.hidden_dim)(x)

2. It gives the illusion that configuration is all bundled in a unique
location, but config is also the preprocessing steps or model layers,
the early-stopping params, the choice of optimizer or loss function,
etc. Those are hard-coded in the notebook. Some could be extracted as
config settings fairly easily, but others are python code. Extracting
everything into a config object would basically mean transforming the
whole tensorflow code into a declarative config (let's not go
there...).

It was better to have a configuration when running `train.py` as a
somewhat static model. But since we're going the notebook route to
facilitate experiments, it's easier to keep the config close to the
code, as most of the notebook is configuration (of the model
structure) anyway.

If we want to tweak the size of hidden dims (or the learning rate or
...), we'll most likely be looking at the cell in which we build the
model, not the top cell with the config. And when we make a change, we
won't have to jump around the notebook to edit the config then go back
and run the affected cell.