Fix NaN bug during training with image embeddings
The current implementation relied on the transformation of NaN into 0
by a custom layer. This transformation is however not differentiable.
All MultiAttentionLayer weights were NaN after the first batch of
training, and it got unnoticed because we removed all NaN after the
GlobalAveragePooling1d layer.

To fix this bug, we always make sure the mask for image embedding is
all zeros (no image).